+++ gandi/cli/tests/commands/test_ip.py
@@ -333,6 +333,40 @@ ip6:\t2001:4b98:dc0:47:216:3eff:feb2:3862""")
         self.assertEqual(params['bandwidth'], 102400)
         self.assertEqual(params['ip_version'], 4)
 
+    def test_create_datacenter_limited(self):
+        args = ['--datacenter', 'FR-SD3', '--bandwidth', '51200',
+                '--ip-version', '6']
+        result = self.invoke_with_exceptions(ip.create, args,
+                                             obj=GandiContextHelper())
+
+        self.assertEqual(re.sub(r'\[#+\]', '[###]',
+                                result.output.strip()), """\
+/!\ Datacenter FR-SD3 will be closed on 25/12/2016, please consider using \
+another datacenter.
+Creating your iface.
+\rProgress: [###] 100.00%  00:00:00  \
+\nYour iface has been created with the following IP addresses:
+ip4:\t95.142.160.181
+ip6:\t2001:4b98:dc0:47:216:3eff:feb2:3862""")
+
+        self.assertEqual(result.exit_code, 0)
+        params = self.api_calls['hosting.iface.create'][0][0]
+        self.assertEqual(params['datacenter_id'], 4)
+        self.assertEqual(params['bandwidth'], 51200)
+        self.assertEqual(params['ip_version'], 6)
+
+    def test_create_datacenter_closed(self):
+        args = ['--datacenter', 'US-BA1', '--bandwidth', '51200',
+                '--ip-version', '6']
+        result = self.invoke_with_exceptions(ip.create, args,
+                                             obj=GandiContextHelper())
+
+        self.assertEqual(re.sub(r'\[#+\]', '[###]',
+                                result.output.strip()), """\
+Error: /!\ Datacenter US-BA1 is closed, please choose another datacenter.""")
+
+        self.assertEqual(result.exit_code, 1)
+
     def test_create_params(self):
         args = ['--datacenter', 'FR', '--bandwidth', '51200',
                 '--ip-version', '6']